Major Cleanup & UI Components

+++ source/D3D11Helper/D3D11TextureHelper.cpp
@@ -85,8 +85,8 @@ TX_IDX CreateTexture(FORMAT format, USAGE_FLAGS useFlags, RESOURCE_FLAGS bindFla
 	uint16_t& current = txHolder->currentCount;
 	D3D11_TEXTURE2D_DESC desc;
 	// Take the height and width of the loaded image and set it as the dimensions for the texture
-	desc.Width = width;
-	desc.Height = height;
+	desc.Width = (UINT)width;
+	desc.Height = (UINT)height;
 	desc.MipLevels = 1;
 	desc.ArraySize = 1;
 	desc.Format = (DXGI_FORMAT)format;// DXGI_FORMAT_R8G8B8A8_UNORM;
